accept_multishot: wrap Stream in io::Result
Handles first part of prior commit:

    See if the method shouldn't support an io::Result instead so an invalid
    operation could be caught before the stream's loop was entered.

Next work: see about removing the try portion of the Stream.

Next work: see if the SocketAddr is correct and important to provide.

See if a remote_addr on the TcpStream wouldn't solve the same purpose.

See about providing a cancel mechanism, through the use of an
AsyncCancel operation.

Outstanding bug: dropping the Stream does not get the operation canceled
so there currently is no way to exit the runtime gracefully. A graceful
runtime shutdown requires no outstanding operations and currently there
is no mechanism to stop the multi shot accept operation once given to
the kernel. That will require a sync cancel or an async cancel operation
be performed and to be tied to the Stream being dropped.

Also would be nice if the caller could stop/cancel the multi operation
themselves, as dropping the Stream while it is blocked on the 'next'
call is not possible.