accept_multishot example, replace try_next with next
The example loop shouldn't abort if a next result from the Stream
is an error. Now just print the error and continue.

Also resolved that the fact the Stream returns the remote SocketAddr
as the second field of the tuple conforms to what the tokio
TcpListner.accept does. Although there it comes from the mio crate
so it passed it on without extra work. Going to leave it in for now
but it does seem like extra work and the resulting TcpStream should
have a remote_addr method on it eventually anyway.

Remaining:

See about providing a cancel mechanism, through the use of an
AsyncCancel operation.

Outstanding bug: dropping the Stream does not get the operation canceled
so there currently is no way to exit the runtime gracefully. A graceful
runtime shutdown requires no outstanding operations and currently there
is no mechanism to stop the multi shot accept operation once given to
the kernel. That will require a sync cancel or an async cancel operation
be performed and to be tied to the Stream being dropped.

Also would be nice if the caller could stop/cancel the multi operation
themselves, as dropping the Stream while it is blocked on the 'next'
call is not possible.