chore: fix race of fs unit tests (#121)
Setting the env variable RUST_TEST_THREADS to 1 in the local
.cargo/config.toml avoids a race during unit testing where some test is
closing a file descriptor and then testing that it is closed while other
tests are opening files.

To manually test the fs commands with the threads override:

    cargo test --test fs_file -- --test-threads=1

With this addition of the .cargo/config.toml, the threads override
does not have to be given manually.

    cargo test --test fs_file

Fixes #119.