Drop SharedFd with sync close, not async operation
A hack or a work-in-progress.

When a tick causes an ignored SharedFd to be dropped, the close can't be
done with the ring close op because the current design would try to
borrow_mut the CURRENT driver, but it's already borrowed mutably to run
tick in the first place.

This removes the user of the ring close op for that and settles on the
synchronous syscall to close the fd instead.